+++ webapp/src/components/ComentContainer/CommentElement/Nested.CommentElement.jsx
@@ -1,12 +1,13 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import { useCommentsAction, useCommentsState } from 'contexts/Comment/Comment.Provider';
+import { useCommentsState } from 'contexts/Comment/Comment.Provider';
 import { replyCommentInfoType } from 'types/comment.type';
 import { parsedNumberToThreeDigits } from 'utils';
 import EditRootCommentForm from '../CommentForm/Edit.CommentForm';
 import SecretCommentElement from './Secret.CommentElement';
 import CommentLikeSvg from './CommentLikeSvg';
 import * as S from '../style';
+import CommentButtonContainer from './CommentButtonContainer';
 
 NestedCommentElement.propTypes = {
   commentInfo: replyCommentInfoType.isRequired,
@@ -14,8 +15,7 @@ NestedCommentElement.propTypes = {
 };
 
 export default function NestedCommentElement({ commentInfo, postWriterId }) {
-  const { editTargetCommentId, postType } = useCommentsState();
-  const { selectEditTargetComment, handleClickDeleteTargetComment } = useCommentsAction();
+  const { editTargetCommentId } = useCommentsState();
   const {
     id: commentId,
     parentId,
@@ -28,10 +28,6 @@ export default function NestedCommentElement({ commentInfo, postWriterId }) {
   const { id: writerId, image: writerProfileImage, name: writerName } = writerInfo;
   const likesCount = likedUserIds.length;
   const isEditTargetComment = commentId === editTargetCommentId;
-
-  const handleClickDeleteButton = () => {
-    handleClickDeleteTargetComment({ postType, id: commentId });
-  };
 
   return (
     <S.NestedCommentBox>
@@ -53,17 +49,12 @@ export default function NestedCommentElement({ commentInfo, postWriterId }) {
           <CommentLikeSvg commentId={commentId} parentId={parentId} likedUserIds={likedUserIds} />
           <span>{parsedNumberToThreeDigits(likesCount)}</span>
         </S.SpecificInfo>
-        {!isEditTargetComment && (
-          <S.SpecificInfo>
-            <S.EditButton isNested onClick={() => selectEditTargetComment(commentId)}>
-              댓글수정
-            </S.EditButton>
-            <span>/</span>
-            <S.EditButton isNested onClick={handleClickDeleteButton}>
-              삭제하기
-            </S.EditButton>
-          </S.SpecificInfo>
-        )}
+        <CommentButtonContainer
+          isNested
+          isEditTargetComment={isEditTargetComment}
+          commentId={commentId}
+          writerId={writerId}
+        />
       </S.CommentInfo>
       {isEditTargetComment && (
         <EditRootCommentForm initialText={content} secret={secret} commentId={commentId} />
